use templates to make it header-only, move to ./src/inc/til/
