+++ src/inc/til.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include "til/some.h"
+#include "til/u8u16convert.h"
 
 namespace til // Terminal Implementation Library. Also: "Today I Learned"
 {
